Switch to use DTO for Onejav
