+++ Modules/Core/app/Dto/Traits/TDefaultDto.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Modules\Core\Dto\Traits;
+
+trait TDefaultDto
+{
+    public function getFields(): array
+    {
+        return [];
+    }
+}
