+++ Modules/Jav/app/Jobs/OnejavFetchItemsJob.php
@@ -8,8 +8,8 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Modules\Core\Services\SettingService;
+use Modules\Jav\Client\Onejav\CrawlingService;
 use Modules\Jav\Repositories\OnejavRepository;
-use Modules\Jav\Services\Onejav\CrawlingService;
 use Modules\Jav\Services\Onejav\OnejavService;
 use Throwable;
