+++ Modules/Jav/tests/Unit/Onejav/CrawlingServiceTest.php
@@ -5,10 +5,11 @@ namespace Modules\Jav\Tests\Unit\Onejav;
 use Illuminate\Support\Facades\Event;
 use Modules\Client\Services\ClientManager;
 use Modules\Jav\Client\Onejav\Client as OnejavClient;
+use Modules\Jav\Client\Onejav\CrawlingService;
+use Modules\Jav\Dto\ItemDto;
 use Modules\Jav\Entities\OnejavItemEntity;
 use Modules\Jav\Events\CrawlingFailedEvent;
 use Modules\Jav\Events\OnejavHaveNextPageEvent;
-use Modules\Jav\Services\Onejav\CrawlingService;
 use Modules\Jav\tests\TestCase;
 
 class CrawlingServiceTest extends TestCase
@@ -24,10 +25,9 @@ class CrawlingServiceTest extends TestCase
 
         $body = $response->parseBody();
 
-        $item = app(CrawlingService::class)
-            ->item($body->getData());
+        $item = app(CrawlingService::class)->item($body->getData());
 
-        $this->assertInstanceOf(OnejavItemEntity::class, $item);
+        $this->assertInstanceOf(ItemDto::class, $item);
         $this->assertEquals('/torrent/cvdx588', $item->url);
         $this->assertEquals('https://pics.dmm.co.jp/mono/movie/adult/h_086cvdx588/h_086cvdx588pl.jpg', $item->cover);
         $this->assertEquals('CVDX-588', $item->dvd_id);
